Fix a crash when restoring flow controller.
